+++ lib/request_handler/builder/base.rb
@@ -1,0 +1,23 @@
+# frozen_string_literal: true
+
+require 'ostruct'
+
+module RequestHandler
+  module Builder
+    class Base
+      attr_accessor :result
+
+      def initialize
+        create_klass_struct
+      end
+
+      def create_klass_struct
+        raise NotImplementedError
+      end
+
+      def build
+        result
+      end
+    end
+  end
+end