+++ lib/request_handler/builder/fieldsets_resource_builder.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require "request_handler/builder/base"
+
+module RequestHandler
+  module Builder
+    class FieldsetsResourceBuilder < Base
+      def create_klass_struct
+        @result = OpenStruct.new
+      end
+
+      def resource(name, value)
+        @result[name.to_sym] = value
+      end
+    end
+  end
+end