+++ spec/integration/fieldsets_parser_spec.rb
@@ -12,7 +12,7 @@ describe RequestHandler do
           options do
             fieldsets do
               allowed do
-                posts(schema)
+                resource :posts, schema
               end
               required [:posts]
             end
@@ -63,7 +63,7 @@ describe RequestHandler do
       end
 
       context 'with a value that is not allowed for a type' do
-        before { testclass.config.fieldsets.allowed.posts = %w[foo bar] }
+        before { testclass.config.config.fieldsets.allowed.posts = %w[foo bar] }
         let(:request) { build_mock_request(params: { 'fields' => { 'posts' => 'foo' } }, headers: nil, body: '') }
         it { expect { to_dto }.to raise_error(RequestHandler::InternalArgumentError) }
       end
@@ -75,7 +75,7 @@ describe RequestHandler do
           options do
             fieldsets do
               allowed do
-                posts true
+                resource :posts, true
               end
             end
           end